refactor: unify hw info page into one component
